Ending backgrounds
0 suspects or less than 5 clues:
You got fired from the police department. No one was arrested for the threats to the goverment and the Lux organisation can continue their operations.
The future will show if the government can still stop the organisation, or if Lux gets the upper hand in the city of Porto.

G:1 I:0 
The Lux organisation has been weakened by the arrest of one of their key members. This will keep the city safe for now, but the threat of the Lux organisation still remains...

G2 I:0
You have arrested both of the key members and disbanded the Lux organisation. The government and also the city of Porto is safe from the threats and will try to comply with some idealogies
of the Lux' progressive ideas.

G:0 I:1-2
You got fired from the police department. The innocent people you have arrested have been released but the people also lost interest in the authorities.
More and more people join the cause of the Lux movement, but the government still has a chance to stop Lux from overthrowing them.

G:0 I:3
You are a fugitive from your own city. People went to riots and Lux gain the upper hand, gaining a lot of attention and also creating a chain reaction from people with 
other idealogies. The city of Porto is in shambles after all the riots and other kind of attacks in the city, while the government is disbanded.

G:X I:Y
You were demoted to paperwork. You may have found some leads for the Lux organisation, but also falsely accused some innocent people. Lux has been weakened by the uncovered information
and the arrest, but the people of Porto also doubt if the government is competent enough to protect them from futures threats and attacks.

All suspects:
You had to flee the country. People were furious and scared of the random arrests and questioned the competence of the police and therefore of the government. Lux gained more followers
and power and the government was disbanded.